* Implement paging library in popular movies list. * Remove PopularMoviesRepository and use PopularMoviesRemoteDataSourceFactory instead.

+++ app/src/main/java/com/mina_mikhail/fixed_solutions_task/data/source/remote/ApiInterface.java
@@ -11,7 +11,7 @@ public interface ApiInterface {
 
   @GET("discover/movie")
   Single<PopularMoviesResponse> getMovies(@Query("sort_by") String sortBy,
-      @Query("page") int pageNumber);
+      @Query("page") long pageNumber);
 
   @GET("movie/{movieID}")
   Single<MovieDetails> getMovieDetails(@Path("movieID") int movieID);
